\name{newview-package}
\alias{newview-package}
\alias{newview}
\docType{package}
\title{\packageTitle{newview}}
\description{\packageDescription{newview}}
\section{Package Content}{\packageIndices{newview}}
\author{\packageAuthor{newview}}
\section{Maintainer}{\packageMaintainer{newview}}
\keyword{package}
